Recomputation for Lmloop matrices

+++ Sparse_CCJ/base.h
@@ -6,7 +6,7 @@
 
 //! type of energy
 // Should not be a short as INF is 1600000 and is a very common value for energy.
-typedef long int energy_t;
+typedef int energy_t;
 typedef unsigned short index_t;
